+++ core/clients/key_flow.go
@@ -36,7 +36,7 @@ const (
 type KeyFlow struct {
 	client        *http.Client
 	config        *KeyFlowConfig
-	doer          func(client *http.Client, req *http.Request, cfg *RetryConfig) (resp *http.Response, err error)
+	doer          func(req *http.Request) (resp *http.Response, err error)
 	key           *ServiceAccountKeyResponse
 	privateKey    *rsa.PrivateKey
 	privateKeyPEM []byte
@@ -47,8 +47,9 @@ type KeyFlow struct {
 
 // KeyFlowConfig is the flow config
 type KeyFlowConfig struct {
-	ServiceAccountKey             *ServiceAccountKeyResponse
-	PrivateKey                    string
+	ServiceAccountKey *ServiceAccountKeyResponse
+	PrivateKey        string
+	// Deprecated: retry options were removed to reduce complexity of the client. If this functionality is needed, you can provide your own custom HTTP client.
 	ClientRetry                   *RetryConfig
 	TokenUrl                      string
 	BackgroundTokenRefreshContext context.Context // Functionality is enabled if this isn't nil
@@ -119,15 +120,11 @@ func (c *KeyFlow) Init(cfg *KeyFlowConfig) error {
 	// No concurrency at this point, so no mutex check needed
 	c.token = &TokenResponseBody{}
 	c.config = cfg
-	c.doer = Do
 
 	if c.config.TokenUrl == "" {
 		c.config.TokenUrl = tokenAPI
 	}
 	c.configureHTTPClient()
-	if c.config.ClientRetry == nil {
-		c.config.ClientRetry = NewRetryConfig()
-	}
 	err := c.validate()
 	if err != nil {
 		return err
@@ -175,11 +172,15 @@ func (c *KeyFlow) RoundTrip(req *http.Request) (*http.Response, error) {
 		return nil, err
 	}
 	req.Header.Set("Authorization", fmt.Sprintf("Bearer %s", accessToken))
-	return c.doer(c.client, req, c.config.ClientRetry)
+	return c.doer(req)
 }
 
 // GetAccessToken returns a short-lived access token and saves the access and refresh tokens in the token field
 func (c *KeyFlow) GetAccessToken() (string, error) {
+	if c.client == nil {
+		return "", fmt.Errorf("nil http client, please run Init()")
+	}
+
 	c.tokenMutex.RLock()
 	accessToken := c.token.AccessToken
 	c.tokenMutex.RUnlock()
@@ -207,6 +208,7 @@ func (c *KeyFlow) configureHTTPClient() {
 	client := &http.Client{}
 	client.Timeout = DefaultClientTimeout
 	c.client = client
+	c.doer = c.client.Do
 }
 
 // validate the client is configured well
@@ -277,6 +279,10 @@ func (c *KeyFlow) createAccessToken() (err error) {
 // createAccessTokenWithRefreshToken creates an access token using
 // an existing pre-validated refresh token
 func (c *KeyFlow) createAccessTokenWithRefreshToken() (err error) {
+	if c.client == nil {
+		return fmt.Errorf("nil http client, please run Init()")
+	}
+
 	c.tokenMutex.RLock()
 	refreshToken := c.token.RefreshToken
 	c.tokenMutex.RUnlock()
@@ -328,7 +334,7 @@ func (c *KeyFlow) requestToken(grant, assertion string) (*http.Response, error) 
 		return nil, err
 	}
 	req.Header.Add("Content-Type", "application/x-www-form-urlencoded")
-	return c.doer(&http.Client{}, req, c.config.ClientRetry)
+	return c.doer(req)
 }
 
 // parseTokenResponse parses the response from the server