+++ core/clients/key_flow_continuous_refresh_test.go
@@ -100,7 +100,7 @@ func TestContinuousRefreshToken(t *testing.T) {
 			}
 
 			numberDoCalls := 0
-			mockDo := func(client *http.Client, req *http.Request, cfg *RetryConfig) (resp *http.Response, err error) {
+			mockDo := func(req *http.Request) (resp *http.Response, err error) {
 				numberDoCalls++
 
 				if tt.doError != nil {
@@ -135,10 +135,10 @@ func TestContinuousRefreshToken(t *testing.T) {
 
 			keyFlow := &KeyFlow{
 				config: &KeyFlowConfig{
-					ClientRetry:                   NewRetryConfig(),
 					BackgroundTokenRefreshContext: ctx,
 				},
-				doer: mockDo,
+				client: &http.Client{},
+				doer:   mockDo,
 				token: &TokenResponseBody{
 					AccessToken:  accessToken,
 					RefreshToken: refreshToken,
@@ -229,7 +229,7 @@ func TestContinuousRefreshTokenConcurrency(t *testing.T) {
 	doTestPhase1RequestDone := false
 	doTestPhase2RequestDone := false
 	doTestPhase4RequestDone := false
-	mockDo := func(client *http.Client, req *http.Request, cfg *RetryConfig) (resp *http.Response, err error) {
+	mockDo := func(req *http.Request) (resp *http.Response, err error) {
 		switch currentTestPhase {
 		default:
 			t.Fatalf("Do call: unexpected request during test phase %d", currentTestPhase)
@@ -330,7 +330,6 @@ func TestContinuousRefreshTokenConcurrency(t *testing.T) {
 	keyFlow := &KeyFlow{
 		client: &http.Client{},
 		config: &KeyFlowConfig{
-			ClientRetry:                   NewRetryConfig(),
 			BackgroundTokenRefreshContext: ctx,
 		},
 		doer: mockDo,