+++ core/CHANGELOG.md
@@ -1,6 +1,36 @@
-## v0.10.0 (YYYY-MM-DD)
+## v0.10.0 (2024-02-27)
 
 - **Feature:** Add configuration option that, for the key flow, enables a goroutine to be spawned that will refresh the access token when it's close to expiring
+- **Deprecation:**
+  - Methods:
+    - `config.WithMaxRetries`
+    - `config.WithWaitBetweenCalls`
+    - `config.WithRetryTimeout`
+    - `clients.NewRetryConfig`
+  - Fields:
+    - `clients.KeyFlowConfig.ClientRetry`
+    - `clients.TokenFlowConfig.ClientRetry`
+    - `clients.NoAuthFlowConfig.ClientRetry`
+    - `clients.RetryConfig`
+  - Retry options were removed to reduce complexity of the clients. If this functionality is needed, you can provide your own custom HTTP client.
+- **Breaking Change:** Change signature of `auth.NoAuth`, which no longer takes `clients.RetryConfig` as argument.
+- **Breaking Change:**
+  - Methods:
+    - `clients.KeyFlow.Clone`
+    - `clients.TokenFlow.Clone`
+    - `clients.NoAuthFlow.Clone`
+    - `clients.Do`
+  - Fields:
+    - `clients.DefaultRetryMaxRetries`
+    - `clients.DefaultRetryWaitBetweenCalls`
+    - `clients.DefaultRetryTimeout`
+  - Constants:
+    - `clients.ClientTimeoutErr`
+    - `clients.ClientContextDeadlineErr`
+    - `clients.ClientConnectionRefusedErr`
+    - `clients.ClientEOFError`
+    - `clients.Environment`
+  - Removed to reduce complexity of the clients, they were no longer being used.
 
 ## v0.9.0 (2024-02-19)